+++ pymodbus/client/tls.py
@@ -26,9 +26,9 @@ class AsyncModbusTlsClient(AsyncModbusTcpClient):
     :param port: Port used for communication
     :param name: Set communication name, used in logging
     :param source_address: Source address of client
-    :param reconnect_delay: Minimum delay before reconnecting, in seconds (float).
-    :param reconnect_delay_max: Maximum delay before reconnecting, in seconds (float).
-    :param timeout: Timeout for connecting and receiving data, in seconds (float).
+    :param reconnect_delay: Minimum delay before reconnecting, in seconds.
+    :param reconnect_delay_max: Maximum delay before reconnecting, in seconds.
+    :param timeout: Timeout for connecting and receiving data, in seconds.
     :param retries: Max number of retries per request.
     :param trace_packet: Called with bytestream received/to be sent
     :param trace_pdu: Called with PDU received/to be sent
@@ -134,7 +134,7 @@ class ModbusTlsClient(ModbusTcpClient):
     :param source_address: Source address of client
     :param reconnect_delay: Not used in the sync client
     :param reconnect_delay_max: Not used in the sync client
-    :param timeout: Timeout for connecting and receiving data, in seconds (float).
+    :param timeout: Timeout for connecting and receiving data, in seconds.
     :param retries: Max number of retries per request.
     :param trace_packet: Called with bytestream received/to be sent
     :param trace_pdu: Called with PDU received/to be sent
